Move `create_dataframe()` to output_format.py
