+++ tests/e2e/alpha/internal/test_fetch_metrics.py
@@ -32,8 +32,8 @@ from neptune_fetcher.alpha.filters import (
     Filter,
 )
 from neptune_fetcher.alpha.internal import identifiers
-from neptune_fetcher.alpha.internal.composition.fetch_metrics import _create_dataframe
 from neptune_fetcher.alpha.internal.context import get_context
+from neptune_fetcher.alpha.internal.output_format import create_dataframe
 from neptune_fetcher.alpha.internal.retrieval.search import fetch_experiment_sys_attrs
 
 
@@ -166,7 +166,7 @@ def create_expected_data(
             limited = filtered[-tail_limit:] if tail_limit is not None else filtered
             rows.extend(limited)
 
-    df = _create_dataframe(
+    df = create_dataframe(
         rows,
         type_suffix_in_column_names=type_suffix_in_column_names,
         include_point_previews=False,
