+++ tests/unit/alpha/test_fetch_metrics.py
@@ -11,11 +11,11 @@ import pandas as pd
 import pytest
 
 from neptune_fetcher.alpha.internal.composition.fetch_metrics import (
-    _create_dataframe,
     _validate_include_time,
     _validate_step_range,
     _validate_tail_limit,
 )
+from neptune_fetcher.alpha.internal.output_format import create_dataframe
 from neptune_fetcher.alpha.internal.retrieval.metrics import FloatPointValue
 
 EXPERIMENTS = 5
@@ -54,7 +54,7 @@ def test_create_flat_dataframe_shape(include_preview):
     float_point_values = list(generate_float_point_values(EXPERIMENTS, PATHS, STEPS, include_preview))
 
     """Test the creation of a flat DataFrame from float point values."""
-    df = _create_dataframe(
+    df = create_dataframe(
         float_point_values,
         include_point_previews=include_preview,
         type_suffix_in_column_names=False,
@@ -97,7 +97,7 @@ def test_create_dataframe_with_absolute_timestamp(type_suffix_in_column_names: b
         ("exp2", "path1", _make_timestamp(2023, 1, 2), 1, 30.0, True, 0.5),
     ]
 
-    df = _create_dataframe(
+    df = create_dataframe(
         data,
         timestamp_column_name="absolute_time",
         type_suffix_in_column_names=type_suffix_in_column_names,
@@ -148,7 +148,7 @@ def test_create_dataframe_without_timestamp(type_suffix_in_column_names: bool, i
         ("exp2", "path1", _make_timestamp(2023, 1, 2), 1, 30.0, True, 0.5),
     ]
 
-    df = _create_dataframe(
+    df = create_dataframe(
         data,
         type_suffix_in_column_names=type_suffix_in_column_names,
         include_point_previews=include_preview,
@@ -188,7 +188,7 @@ def test_create_empty_dataframe(type_suffix_in_column_names: bool, include_previ
     # Given empty dataframe
 
     # When
-    df = _create_dataframe(
+    df = create_dataframe(
         [],
         type_suffix_in_column_names=type_suffix_in_column_names,
         include_point_previews=include_preview,
@@ -230,7 +230,7 @@ def test_create_dataframe_with_reserved_paths_with_multiindex(
         ("exp2", path, _make_timestamp(2023, 1, 2), 1, 30.0, True, 0.5),
     ]
 
-    df = _create_dataframe(
+    df = create_dataframe(
         data,
         timestamp_column_name=timestamp_column_name,
         type_suffix_in_column_names=type_suffix_in_column_names,
@@ -287,7 +287,7 @@ def test_create_dataframe_with_reserved_paths_with_flat_index(path, type_suffix_
         ("exp2", path, _make_timestamp(2023, 1, 2), 1, 30.0, True, 0.5),
     ]
 
-    df = _create_dataframe(
+    df = create_dataframe(
         data,
         type_suffix_in_column_names=type_suffix_in_column_names,
         include_point_previews=False,
